+++ packages/ui/src/components/Routes/Preview.stories.tsx
@@ -1,12 +1,10 @@
 import {
   FrameQuery,
   Locale,
-  OperationExecutorsProvider,
   PreviewDrupalPageQuery,
   Url,
 } from '@custom/schema';
 import { Meta, StoryObj } from '@storybook/react';
-import React from 'react';
 
 import { AccordionItemText } from '../Organisms/PageContent/BlockAccordion.stories';
 import { Default as BlockImageTeasers } from '../Organisms/PageContent/BlockImageTeasers.stories';
@@ -24,67 +22,63 @@ export default {
 } satisfies Meta<typeof Preview>;
 
 export const Default = {
-  render: (args) => {
-    return (
-      <OperationExecutorsProvider
-        executors={[
-          { executor: args, id: PreviewDrupalPageQuery },
-          { id: FrameQuery, executor: FrameStory.args },
-        ]}
-      >
-        <Preview />
-      </OperationExecutorsProvider>
-    );
-  },
-  args: {
-    preview: {
-      title: 'Preview Page Title',
-      locale: 'en',
-      translations: [
-        {
-          locale: Locale.En,
-          path: '/test' as Url,
-        },
-      ],
-      path: '/test' as Url,
-      content: [
-        {
-          __typename: 'BlockImageTeasers',
-          ...BlockImageTeasers.args,
-        },
-        {
-          __typename: 'BlockMarkup',
-          ...Mixed.args,
-        },
-        {
-          __typename: 'BlockMedia',
-          ...WithCaption.args,
-        },
-        {
-          __typename: 'BlockMarkup',
-          ...Paragraph.args,
-        },
-        {
-          __typename: 'BlockImageWithText',
-          ...ImageRight.args,
-        },
-        {
-          __typename: 'BlockAccordion',
-          ...AccordionItemText.args,
-        },
-        {
-          __typename: 'BlockImageTeasers',
-          ...BlockImageTeasers.args,
-        },
-      ] as Exclude<PreviewDrupalPageQuery['preview'], undefined>['content'],
-    },
-  },
   parameters: {
+    executors: {
+      [PreviewDrupalPageQuery]: {
+        preview: {
+          title: 'Preview Page Title',
+          locale: 'en',
+          translations: [
+            {
+              locale: Locale.En,
+              path: '/test' as Url,
+            },
+          ],
+          path: '/test' as Url,
+          content: [
+            {
+              __typename: 'BlockImageTeasers',
+              ...BlockImageTeasers.args,
+            },
+            {
+              __typename: 'BlockMarkup',
+              ...Mixed.args,
+            },
+            {
+              __typename: 'BlockMedia',
+              ...WithCaption.args,
+            },
+            {
+              __typename: 'BlockMarkup',
+              ...Paragraph.args,
+            },
+            {
+              __typename: 'BlockImageWithText',
+              ...ImageRight.args,
+            },
+            {
+              __typename: 'BlockAccordion',
+              ...AccordionItemText.args,
+            },
+            {
+              __typename: 'BlockImageTeasers',
+              ...BlockImageTeasers.args,
+            },
+          ] as Exclude<PreviewDrupalPageQuery['preview'], undefined>['content'],
+        },
+      },
+      [FrameQuery]: FrameStory.parameters.executors[FrameQuery],
+    },
     location: new URL('local:/gatsby-turbo'),
   },
 } satisfies StoryObj<PreviewDrupalPageQuery>;
 
 export const Preview403 = {
-  ...Default,
-  args: {},
+  parameters: {
+    executors: {
+      [PreviewDrupalPageQuery]: {},
+      [FrameQuery]: FrameStory.parameters.executors[FrameQuery],
+    },
+    location: new URL('local:/gatsby-turbo'),
+  },
 } satisfies StoryObj<PreviewDrupalPageQuery>;